+++ native/src/resource.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define IDI_SYMBION 101